#408 renamed from chat-api to chatbot-api
